chore: heartbeat
add heartbeat method for github actions lambda function
