+++ lambda/actions/src/actions/createEvent.ts
@@ -1,0 +1,45 @@
+import { models } from '../../../shared/sequelize/models/models';
+
+const createEvent = async (data) => {
+  try {
+    const result = await models.event.create(data);
+  } catch (err) {
+    console.error(err);
+  }
+};
+
+export default async function status(event) {
+  const { body, queryStringParameters } = event;
+  const { prNumber, prSuccess, planSuccess, applySuccess, id, actionNumber, planDetails } = JSON.parse(body);
+  const { status: githubActionsStage } = queryStringParameters || {};
+
+  if (githubActionsStage === 'create') {
+    const status = String(prSuccess) === 'true' ? 'pr' : 'prFailed';
+    await Promise.all([
+      models.request.update({ prNumber, status, actionNumber }, { where: { id } }),
+      createEvent({ eventCode: `request-pr-${status}`, requestId: id }),
+    ]);
+  } else {
+    // After creation, gh action only has prNumber to reference request. Using this to grab the requestId first
+    const request = await models.request.findOne({ where: { prNumber } });
+    if (!request) throw Error(`request associated with pr number ${prNumber} not found`);
+
+    const { id: requestId, status: currentStatus } = request;
+    const isAlreadyApplied = currentStatus === 'applied';
+    if (githubActionsStage === 'plan') {
+      const status = String(planSuccess) === 'true' ? 'planned' : 'planFailed';
+      await Promise.all([
+        !isAlreadyApplied && models.request.update({ status }, { where: { id: requestId } }),
+        createEvent({ eventCode: `request-plan-${status}`, requestId, planDetails }),
+      ]);
+    }
+    if (githubActionsStage === 'apply') {
+      const status = String(applySuccess) === 'true' ? 'applied' : 'applyFailed';
+      await Promise.all([
+        models.request.update({ status }, { where: { id: requestId } }),
+        createEvent({ eventCode: `request-apply-${status}`, requestId }),
+      ]);
+    }
+  }
+  return {};
+}
