+++ lambda/actions/src/main.ts
@@ -1,12 +1,13 @@
 import { APIGatewayProxyEvent, Context, Callback } from 'aws-lambda';
-import { models, sequelize } from '../../shared/sequelize/models/models';
 import { Response } from './interfaces';
+import { wakeUpAll } from './actions/wakeUp';
+import createEvent from './actions/createEvent';
 
 const responseHeaders = {
   'Content-Type': 'text/html; charset=utf-8',
   'Access-Control-Allow-Headers': 'Content-Type, Authorization',
   'Access-Control-Allow-Origin': '*',
-  'Access-Control-Allow-Methods': 'OPTIONS,PUT',
+  'Access-Control-Allow-Methods': 'OPTIONS,PUT,GET',
 };
 
 const response = (data?: Response) => {
@@ -27,56 +28,26 @@ const unprocessableEntityError = (body = 'unable to process the request') => {
   return response({ statusCode: 422, body: JSON.stringify(body) });
 };
 
-const createEvent = async (data) => {
-  try {
-    const result = await models.event.create(data);
-  } catch (err) {
-    console.error(err);
-  }
-};
-
 export const handler = async (event: APIGatewayProxyEvent, context?: Context, callback?: Callback) => {
   // Sequelize waits ~10seconds to drop connection, delaying API response.
   // Use this to prevent, see https://forum.serverless.com/t/lambda-with-rds-using-vpc-works-slow/1261/7 for more
   context.callbackWaitsForEmptyEventLoop = false;
-  const { headers, body, queryStringParameters } = event;
+  const { headers, body, path, requestContext } = event;
+  const { httpMethod } = requestContext;
   console.log('event.body', body);
-  const { prNumber, prSuccess, planSuccess, applySuccess, id, actionNumber, planDetails } = JSON.parse(body);
+
   const { Authorization } = headers;
   if (Authorization !== process.env.GH_SECRET) return callback(null, unauthorizedError());
 
-  const { status: githubActionsStage } = queryStringParameters || {};
-
   try {
-    if (githubActionsStage === 'create') {
-      const status = String(prSuccess) === 'true' ? 'pr' : 'prFailed';
-      await Promise.all([
-        models.request.update({ prNumber, status, actionNumber }, { where: { id } }),
-        createEvent({ eventCode: `request-pr-${prSuccess}`, requestId: id }),
-      ]);
-    } else {
-      // After creation, gh action only has prNumber to reference request. Using this to grab the requestId first
-      const request = await models.request.findOne({ where: { prNumber } });
-      if (!request) throw Error(`request associated with pr number ${prNumber} not found`);
-
-      const { id: requestId, status: currentStatus } = request;
-      const isAlreadyApplied = currentStatus === 'applied';
-      if (githubActionsStage === 'plan') {
-        const status = String(planSuccess) === 'true' ? 'planned' : 'planFailed';
-        await Promise.all([
-          !isAlreadyApplied && models.request.update({ status }, { where: { id: requestId } }),
-          createEvent({ eventCode: `request-plan-${planSuccess}`, requestId, planDetails }),
-        ]);
-      }
-      if (githubActionsStage === 'apply') {
-        const status = String(applySuccess) === 'true' ? 'applied' : 'applyFailed';
-        await Promise.all([
-          models.request.update({ status }, { where: { id: requestId } }),
-          createEvent({ eventCode: `request-apply-${planSuccess}`, requestId }),
-        ]);
-      }
+    if (httpMethod === 'PUT') {
+      const result = await createEvent(event);
+      return callback(null, response(result));
     }
-    callback(null, response());
+    if (httpMethod === 'GET') {
+      const result = await wakeUpAll();
+      return callback(null, response(result));
+    }
   } catch (err) {
     console.error(err);
     callback(null, unprocessableEntityError(err.message || err));
